+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - add image format selection to `modify_images`
 
+- add automatic bookmark detection to `update_bookmarks`
+
 ### Fixed
 
 - `subset_fonts`: add subset tags to font names